// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  name: any;
  uuid: any;
}
// ...
export interface Tasks {
  __typename?: 'tasks';
  authorId: Scalars['uuid'];
  description: Scalars['String'];
  title: Scalars['name'];
  /** An object relationship */
  user: Users;
  uuid: Scalars['uuid'];
}
// ...
export interface Users {
  __typename?: 'users';
  displayName: Scalars['name'];
  uuid: Scalars['uuid'];
}
// ...
export type TasksQueryVariables = Exact<{ [key: string]: never; }>;


export type TasksQuery = (
  { __typename?: 'query_root' }
  & { tasks: Array<(
    { __typename?: 'tasks' }
    & Pick<Tasks, 'uuid' | 'title' | 'description'>
    & { user: (
      { __typename?: 'users' }
      & Pick<Users, 'displayName' | 'uuid'>
    ) }
  )> }
);

export type AddTaskMutationVariables = Exact<{
  authorId: Scalars['uuid'];
  description: Scalars['String'];
  title: Scalars['name'];
}>;


export type AddTaskMutation = (
  { __typename?: 'mutation_root' }
  & { insert_tasks?: Maybe<(
    { __typename?: 'tasks_mutation_response' }
    & { returning: Array<(
      { __typename?: 'tasks' }
      & Pick<Tasks, 'description' | 'title' | 'uuid'>
      & { user: (
        { __typename?: 'users' }
        & Pick<Users, 'displayName' | 'uuid'>
      ) }
    )> }
  )> }
);

export type TasksSubscriptionSubscriptionVariables = Exact<{ [key: string]: never; }>;


export type TasksSubscriptionSubscription = (
  { __typename?: 'subscription_root' }
  & { tasks: Array<(
    { __typename?: 'tasks' }
    & Pick<Tasks, 'authorId' | 'description' | 'title'>
  )> }
);

export const TasksDocument = gql`
    query Tasks {
  tasks {
    uuid
    title
    description
    user {
      displayName
      uuid
    }
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class TasksGQL extends Apollo.Query<TasksQuery, TasksQueryVariables> {
    document = TasksDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const AddTaskDocument = gql`
    mutation AddTask($authorId: uuid!, $description: String!, $title: name!) {
  insert_tasks(
    objects: {authorId: $authorId, description: $description, title: $title}
  ) {
    returning {
      description
      title
      user {
        displayName
        uuid
      }
      uuid
    }
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class AddTaskGQL extends Apollo.Mutation<AddTaskMutation, AddTaskMutationVariables> {
    document = AddTaskDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }